Fixes #38446 - Move plugin registration to own file
